chore: data context utils for server
